requirements.txt
bedrock-agentcore
strands-agents
strands-agents-tools
boto3
redis
pydantic
pydantic-settings
httpx
structlog
